Refactor namespace imports in repository and controller files for consistency
