+++ CarSpot.Application/Interfaces/IBusinessRepository.cs
@@ -1,9 +1,6 @@
 using CarSpot.Domain.Entities;
-using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 
-namespace CarSpot.Application.Interfaces.Repositories
+namespace CarSpot.Application.Interfaces
 {
     public interface IBusinessRepository
     {
