+++ CarSpot.WebApi/Controllers/BusinessController.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using CarSpot.Application.DTOs;
-using CarSpot.Application.Interfaces.Repositories;
+using CarSpot.Application.Interfaces;
 using CarSpot.Domain.Entities;
 using Microsoft.AspNetCore.Mvc;
